+++ tests/testthat/test-esem.from.mods.R
@@ -2,7 +2,7 @@ test_that(
   "Test normal behaviour: EFA and CFA only",
   {
     esem_fit <- esem.from.mods(
-      efa_fit, cfa_fit, d = BFIGritHope, fit_save = FALSE, check = FALSE
+      efa_fit, cfa_fit, data = BFIGritHope, fit_save = FALSE
     )
     expect_equal(length(esem_fit), 5)
     expect_equal(length(esem_fit$fit), length(keys))
@@ -18,8 +18,7 @@ test_that(
   "Test normal behaviour: EFA and bifactor only",
   {
     esem_fit <- esem.from.mods(
-      efa_fit, bif_fit = bif_fit, d = BFIGritHope,
-      fit_save = FALSE, check = FALSE
+      efa_fit, bif_fit = bif_fit, data = BFIGritHope, fit_save = FALSE
     )
     expect_equal(length(esem_fit), 5)
     expect_equal(length(esem_fit$fit), length(keys_g))
@@ -35,8 +34,8 @@ test_that(
   "Test normal behaviour: EFA and both CFA and bifactor",
   {
     esem_fit <- esem.from.mods(
-      efa_fit, cfa_fit = cfa_fit, bif_fit = bif_fit, d = BFIGritHope,
-      fit_save = TRUE, check = FALSE
+      efa_fit, cfa_fit = cfa_fit, bif_fit = bif_fit, data = BFIGritHope,
+      fit_save = TRUE
     )
     expect_equal(length(esem_fit), 5)
     expect_equal(length(esem_fit$fit), length(keys_g) + length(keys))
@@ -53,7 +52,7 @@ test_that(
   "Neither cfa_fit nor bif_fit set",
   {
     expect_error(
-      esem.from.mods(efa_fit, d = BFIGritHope, fit_save = FALSE, check = FALSE),
+      esem.from.mods(efa_fit, data = BFIGritHope, fit_save = FALSE),
       "one of `cfa_fit` and `bif_fit` must be specified"
     )
   }
@@ -63,22 +62,22 @@ test_that(
   {
     expect_error(
       esem.from.mods(
-        efa_fit, cfa_fit = list(a = 1:2, b = 3:4), d = BFIGritHope,
-        fit_save = FALSE, check = FALSE
+        efa_fit, cfa_fit = list(a = 1:2, b = 3:4), data = BFIGritHope,
+        fit_save = FALSE
       ),
       "`cfa_fit` are not objects of type lavaan"
     )
     expect_error(
       esem.from.mods(
-        efa_fit, bif_fit = list(a = 1:2, b = 3:4), d = BFIGritHope,
-        fit_save = FALSE, check = FALSE
+        efa_fit, bif_fit = list(a = 1:2, b = 3:4), data = BFIGritHope,
+        fit_save = FALSE
       ),
       "`bif_fit` are not objects of type lavaan"
     )
     expect_error(
       esem.from.mods(
-        efa_fit = list(a = 1:2, b = 3:4), cfa_fit = cfa_fit, d = BFIGritHope,
-        fit_save = FALSE, check = FALSE
+        efa_fit = list(a = 1:2, b = 3:4), cfa_fit = cfa_fit, data = BFIGritHope,
+        fit_save = FALSE
       ),
       "`efa_fit` is not an object of type lavaan"
     )
@@ -89,15 +88,15 @@ test_that(
   {
     expect_warning(
       esem.from.mods(
-        efa_fit, setNames(cfa_fit, nm = letters[1:4]), d = BFIGritHope,
-        fit_save = FALSE, check = FALSE
+        efa_fit, setNames(cfa_fit, nm = letters[1:4]), data = BFIGritHope,
+        fit_save = FALSE
       ),
       "names of `cfa_fit` do not match the factor names"
     )
     expect_warning(
       esem.from.mods(
         efa_fit, bif_fit = setNames(bif_fit, nm = letters[1:2]),
-        d = BFIGritHope, fit_save = FALSE, check = FALSE
+        data = BFIGritHope, fit_save = FALSE
       ),
       "names of `bif_fit` do not match the general factor names"
     )
@@ -108,8 +107,7 @@ test_that(
   {
     expect_error(
       esem.from.mods(
-        efa_fit = NULL, cfa_fit, d = BFIGritHope,
-        fit_save = FALSE, check = FALSE
+        efa_fit = NULL, cfa_fit, data = BFIGritHope, fit_save = FALSE
       ),
       "`efa_fit` is NULL"
     )
@@ -128,7 +126,7 @@ test_that(
     cfa2_fit <- list(grit = lavaan::cfa(cfa2_mod, BFIGritHope))
     expect_error(
       esem.from.mods(
-        efa_fit, cfa2_fit, d = BFIGritHope, fit_save = FALSE, check = FALSE
+        efa_fit, cfa2_fit, data = BFIGritHope, fit_save = FALSE
       ),
       "CFA containing more than one latent variable has been found"
     )
@@ -139,15 +137,14 @@ test_that(
   {
     expect_error(
       esem.from.mods(
-        efa_fit, cfa_fit[c(1, 1:4)], d = BFIGritHope,
-        fit_save = FALSE, check = FALSE
+        efa_fit, cfa_fit[c(1, 1:4)], data = BFIGritHope, fit_save = FALSE
       ),
       "two different models in `cfa_fit` have factors with the same name"
     )
     expect_error(
       esem.from.mods(
-        efa_fit, bif_fit = bif_fit[c(1, 1:2)], d = BFIGritHope,
-        fit_save = fit_save, check = FALSE
+        efa_fit, bif_fit = bif_fit[c(1, 1:2)], data = BFIGritHope,
+        fit_save = fit_save
       ),
       "different models in `bif_fit` have general factors with the same name"
     )
@@ -158,11 +155,9 @@ test_that(
   {
     keys2 <- keys
     names(keys2)[1] <- "grit"
-    cfa_fit2 <- cfa.from.keys(
-      keys2, BFIGritHope, fit_save = FALSE, check = FALSE
-    )$fit
+    cfa_fit2 <- cfa.from.keys(keys2, BFIGritHope, fit_save = FALSE)$fit
     expect_error(
-      esem.from.mods(efa_fit, cfa_fit2, bif_fit, BFIGritHope, check = FALSE),
+      esem.from.mods(efa_fit, cfa_fit2, bif_fit, BFIGritHope),
       "models in `cfa_fit` have identically named factor"
     )
   }
@@ -176,39 +171,36 @@ test_that(
     keys_g2$grit <- keys_g2$grit[-1]
     bif_fit2 <- suppressWarnings(  # In preamble, not important for the test.
       bifactor.from.keys(
-        keys_g2, keys_b, keys, BFIGritHope,
-        check = FALSE, fit_save = FALSE
+        keys_g2, keys_b, keys, BFIGritHope, fit_save = FALSE
       )$fit
     )
     expect_no_error(
-      esem.from.mods(
-        efa_fit, bif_fit = bif_fit2, d = BFIGritHope, check = FALSE
-      )
+      esem.from.mods(efa_fit, bif_fit = bif_fit2, data = BFIGritHope)
     )
   }
 )
 test_that(
   "Test `save_out = TRUE` file creation and `check = TRUE` correctly loading",
   {
-    out_dir <- withr::local_tempdir(tmpdir = "tests/testthat")
+    cache_dir <- cache.setup("tests/testthat")
     name <- "esem"
     check_fit <- esem.from.mods(
-      efa_fit, cfa_fit, d = BFIGritHope, check = TRUE, save_out = TRUE,
-      fit_save = TRUE, name = name, out_dir = out_dir
+      efa_fit, cfa_fit, data = BFIGritHope, check = TRUE, save_out = TRUE,
+      fit_save = TRUE, name = name
     )
     expect_all_true(
       c(
-        file.exists(file.path(out_dir, name, paste0(name, "_fit.rds"))),
-        file.exists(file.path(out_dir, name, paste0(name, "_par_std.rds"))),
-        file.exists(file.path(out_dir, name, paste0(name, "_fit_m.rds"))),
-        file.exists(file.path(out_dir, name, paste0(name, "_mod.rds"))),
-        file.exists(file.path(out_dir, name, paste0(name, "_hash.rds")))
+        file.exists(file.path(cache_dir, name, paste0(name, "_fit.rds"))),
+        file.exists(file.path(cache_dir, name, paste0(name, "_par_std.rds"))),
+        file.exists(file.path(cache_dir, name, paste0(name, "_fit_m.rds"))),
+        file.exists(file.path(cache_dir, name, paste0(name, "_mod.rds"))),
+        file.exists(file.path(cache_dir, name, paste0(name, "_hash.rds")))
       )
     )
     check_fit2 <- expect_no_message(
       esem.from.mods(
-        efa_fit, cfa_fit, d = BFIGritHope, check = TRUE, save_out = TRUE,
-        fit_save = TRUE, name = name, out_dir = out_dir
+        efa_fit, cfa_fit, data = BFIGritHope, check = TRUE, save_out = TRUE,
+        fit_save = TRUE, name = name
       ),
       message = "\\d / \\d"
     )