+++ NEWS.md
@@ -1,3 +1,8 @@
+# semFromKeys 0.1.1
+
+* Added caching with [cache.setup()] and [cache.clean()] for use with
+`save_out = TRUE` and `check = TRUE`.
+
 # semFromKeys 0.1.0
 
-* Initial CRAN submission.
+* Initial github version.